Ajuste de venta producto numerado.

+++ dipalza/src/main/java/cl/eos/dipalza/service/resultados/NumeracionResultado.java
@@ -1,11 +1,10 @@
 package cl.eos.dipalza.service.resultados;
 
-import java.math.BigDecimal;
 import java.util.List;
 
 public record NumeracionResultado(
-        Integer cantidadPiezasAsignadas,
-        Integer cantidadPiezasFaltantes,
+        Float cantidadPiezasAsignadas,
+        Float cantidadPiezasFaltantes,
         List<String> numerosPiezasAsignadas,
-        BigDecimal pesoRealDeVenta
+        Float pesoRealDeVenta
 ) {}
